Changed progVersion to 4.1
The records were not being displayed properly.
- Some fields were not being displayed.
- There were errors in the parsing.

Changed variable name for better clarity.
- recordID --> recordVer
- Intended to display the version of the record found.

Updated the debug output for the new variable name.

Updated the user output for the new variable name.

Added recordID to the debug output

Added roleStr to the record parsing
- Was blank in the debug output
- Was included as part of the remainingStr

Added recordCount to count the number of records found.

Added the progName variable.
- Added to debug output
- Added to user output